Add license headers and move helpers to directory above
