--- src/go/configgenerator/clustergen/helpers/doc.go
@@ -1,2 +1,0 @@
-// Package helpers contains utility structs that operate on any generic cluster.
-package helpers
