+++ src/go/configgenerator/clustergen/local_backend.go
@@ -1,7 +1,21 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package clustergen
 
 import (
-	"github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/clustergen/helpers"
+	helpers2 "github.com/GoogleCloudPlatform/esp-v2/src/go/configgenerator/helpers"
 	"github.com/GoogleCloudPlatform/esp-v2/src/go/options"
 	clusterpb "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	scpb "google.golang.org/genproto/googleapis/api/serviceconfig"
@@ -10,8 +24,8 @@ import (
 // LocalBackendCluster is an Envoy cluster to communicate with a local backend
 // that speaks HTTP (OpenAPI) or gRPC (service config) protocol.
 type LocalBackendCluster struct {
-	BackendCluster *helpers.BackendCluster
-	GRPCHealth     *helpers.ClusterGRPCHealthCheckConfiger
+	BackendCluster *helpers2.BackendCluster
+	GRPCHealth     *helpers2.ClusterGRPCHealthCheckConfiger
 }
 
 // NewLocalBackendClusterFromServiceConfig creates a LocalBackendCluster from
@@ -33,7 +47,7 @@ func (c *LocalBackendCluster) GenConfig() (*clusterpb.Cluster, error) {
 		return nil, err
 	}
 
-	if err := helpers.MaybeAddGRPCHealthCheck(c.GRPCHealth, config); err != nil {
+	if err := helpers2.MaybeAddGRPCHealthCheck(c.GRPCHealth, config); err != nil {
 		return nil, err
 	}
