+++ src/go/configgenerator/helpers/tls.go
@@ -1,3 +1,17 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package helpers
 
 import (
@@ -53,7 +67,7 @@ func (c *ClusterTLSConfiger) MakeTLSConfig(hostname string, alpnProtocols []stri
 	//}
 
 	return &corepb.TransportSocket{
-		Name: TLSTransportSockerName,
+		Name:       TLSTransportSockerName,
 		ConfigType: &corepb.TransportSocket_TypedConfig{
 			//TypedConfig: tlsContext,
 		},
